# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CommentAuthorDTO:
    properties:
      admin:
        type: boolean
      avatar:
        type: string
      id:
        type: integer
      nickname:
        type: string
    type: object
  dto.CommentCreateInput:
    properties:
      content:
        type: string
      parent_id:
        type: integer
    type: object
  dto.CommentDTO:
    properties:
      author:
        $ref: '#/definitions/dto.CommentAuthorDTO'
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      parent_id:
        type: integer
      post_id:
        type: integer
      reply_to_user:
        $ref: '#/definitions/dto.CommentAuthorDTO'
      root_id:
        type: integer
      updated_at:
        type: string
    type: object
  dto.CommentListResult:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.CommentDTO'
        type: array
      total:
        type: integer
    type: object
  dto.CommentUpdateInput:
    properties:
      content:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 这是一个轻量级的说说服务
  title: EasyDrop API
  version: "1.0"
paths:
...
  /comments:
    get:
      description: 分页查询公开评论列表
      parameters:
      - description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - description: 每页条数，最大 100
        in: query
        name: size
        type: integer
      - description: 排序，如 created_at_desc 或 created_at_asc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommentListResult'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 前端查询评论列表
      tags:
      - comment
...
  /posts/{id}/comments:
    get:
      description: 分页查询指定说说下的评论（扁平）
      parameters:
      - description: 说说ID
        in: path
        name: id
        required: true
        type: integer
      - description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - description: 每页条数，最大 100
        in: query
        name: size
        type: integer
      - description: 排序，如 created_at_asc 或 created_at_desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommentListResult'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: 说说不存在
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: 前端查询说说评论列表
      tags:
      - comment
    post:
      consumes:
      - application/json
      description: 当前登录用户创建评论
      parameters:
      - description: 说说ID
        in: path
        name: id
        required: true
        type: integer
      - description: 评论创建参数
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CommentCreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommentDTO'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: 说说或用户不存在
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 用户创建评论
      tags:
      - comment
...
  /users/me/comments:
    get:
      description: 分页查询当前登录用户自己的评论列表
      parameters:
      - description: 说说ID
        in: query
        name: post_id
        type: integer
      - description: 页码，从 1 开始
        in: query
        name: page
        type: integer
      - description: 每页条数，最大 100
        in: query
        name: size
        type: integer
      - description: 排序，如 created_at_desc 或 created_at_asc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommentListResult'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: 说说不存在
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 用户查询评论列表
      tags:
      - comment
  /users/me/comments/{id}:
    delete:
      description: 当前登录用户按评论 ID 删除自己的评论
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: 评论不存在
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 用户删除评论
      tags:
      - comment
    get:
      description: 当前登录用户按评论 ID 查询自己的评论详情
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommentDTO'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: 评论不存在
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 用户查询评论详情
      tags:
      - comment
    patch:
      consumes:
      - application/json
      description: 当前登录用户按评论 ID 更新自己的评论
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      - description: 评论更新参数
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CommentUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommentDTO'
        "400":
          description: 参数校验失败
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: 未登录或登录失效
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: 评论不存在
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: 服务内部错误
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 用户更新评论
      tags:
      - comment
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"