# internal/dto/comment.go
package dto

import "time"

type CommentCreateInput struct {
	PostID        uint   `json:"-"`
	UserID        uint   `json:"-"`
	CanViewHidden bool   `json:"-"`
	Content       string `json:"content"`
	ParentID      *uint  `json:"parent_id"`
}

type CommentIDURIInput struct {
	ID uint `uri:"id" binding:"required,gt=0"`
}

type CommentUpdateInput struct {
	ID      uint    `json:"-"`
	Content *string `json:"content"`
}

type CommentListInput struct {
	PostID        uint   `json:"post_id"`
	CanViewHidden bool   `json:"-"`
	Page          int    `json:"page"`
	Size          int    `json:"size"`
	Order         string `json:"order"`
}

type CommentPostListQueryInput struct {
	Page  int    `json:"page" form:"page"`
	Size  int    `json:"size" form:"size"`
	Order string `json:"order" form:"order"`
}

type CommentPublicListQueryInput struct {
	Page  int    `json:"page" form:"page"`
	Size  int    `json:"size" form:"size"`
	Order string `json:"order" form:"order"`
}

type CommentPublicListInput struct {
	CanViewHidden bool   `json:"-"`
	Page          int    `json:"page"`
	Size          int    `json:"size"`
	Order         string `json:"order"`
}

type CommentUserListInput struct {
	UserID        uint   `json:"user_id"`
	PostID        *uint  `json:"post_id" form:"post_id" binding:"omitempty,gt=0"`
	CanViewHidden bool   `json:"-"`
	Page          int    `json:"page" form:"page"`
	Size          int    `json:"size" form:"size"`
	Order         string `json:"order" form:"order"`
}

type CommentAdminListInput struct {
	PostID *uint  `json:"post_id" form:"post_id" binding:"omitempty,gt=0"`
	UserID *uint  `json:"user_id" form:"user_id" binding:"omitempty,gt=0"`
	Page   int    `json:"page" form:"page"`
	Size   int    `json:"size" form:"size"`
	Order  string `json:"order" form:"order"`
}

type CommentDTO struct {
	ID          uint              `json:"id"`
	PostID      uint              `json:"post_id"`
	Author      CommentAuthorDTO  `json:"author"`
	Content     string            `json:"content"`
	ParentID    *uint             `json:"parent_id"`
	RootID      *uint             `json:"root_id"`
	ReplyToUser *CommentAuthorDTO `json:"reply_to_user"`
	CreatedAt   time.Time         `json:"created_at"`
	UpdatedAt   time.Time         `json:"updated_at"`
}

type CommentAuthorDTO struct {
	ID       uint    `json:"id"`
	Nickname string  `json:"nickname"`
	Avatar   *string `json:"avatar"`
	Admin    bool    `json:"admin"`
}

type CommentListResult struct {
	Items []CommentDTO `json:"items"`
	Total int64        `json:"total"`
}